nijis_barcode_system/requirements.txt
requests
flask
brother_ql
pillow
python-barcode